backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlmodel==0.0.37
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
psycopg2-binary==2.9.10
python-dotenv==1.2.2
pydantic-settings==2.12.0
email-validator==2.3.0
boto3==1.40.61
alembic==1.16.5
pytest==8.4.2
httpx==0.28.1
redis==6.2.0
